update and delete privacy for contacts

+++ controllers/contactControllers.js
@@ -125,6 +125,11 @@ const updateContact = asyncHandler(async (req, res) => {
          res.status(404);
          throw new Error("Contact not found!");
     }
+    if (contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("You are not authorized to update this contact");
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -145,6 +150,10 @@ const deleteContact = asyncHandler(async (req, res) => {
     if (!contact) {
          res.status(404);
          throw new Error("Contact not found!");
+    }
+    if (contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("You are not authorized to update this contact");
     }
     await Contact.findByIdAndDelete(req.params.id); // Delete the contact
     res.status(200).json({ message: "Contact deleted successfully" }); // Send a success response
